+++ .env
@@ -1,4 +1,4 @@
 FLASK_ENV=development
 FLASK_DEBUG=1
 FLASK_APP=server_config.py
-MONGO_URI=<URI>
+MONGO_URI=
