+++ src/tints/image/input/input.txt
