+++ src/tints/image/output/output.txt
