+++ src/tints/settings.py
@@ -1,0 +1,5 @@
+#settings.py
+import os
+# __file__ refers to the file settings.py 
+APP_ROOT = os.path.dirname(os.path.abspath(__file__))   # refers to application_top
+APP_INPUT = os.path.join(APP_ROOT, './image/input')
